Update patch set 3
Patch Set 3:

This is the first chunk of converting to the new BSON library. The remaining places where it's used are tests and in places where it isn't easy to change it out. I'm going to start doing some API changes to start returning slices of bytes from functions and methods, so that we can remove the rest of the old BSON library.

Patch-set: 3
Label: Code-Review=0