Update patch set 5
Patch Set 5:

(19 comments)

Responses below. I am concerned at the suggestion that we should only do half of the design. The topology design isn't not fully thought through, per your suggestions I didn't reimplement or attempt to reimplement SDAM and instead chose to design the types in a way that they are compact and still provide the same functionality as before without exposing internals, such as monitors.

I don't understand the suggestion to "modifying connections only as needed to get them to take/give WireMessage, etc." since the connection package right now mostly does encoding and decoding. The connection package as designed is more or less a stripped down version of the current connection package with the additional types for speaking the wire protocol in both directions.

Additionally, it isn't a good idea to omit the receiving parts of the connection package and throw them under out of scope unless we are saying that we think tools that need this (atlasproxy, mongoreply, etc...) should use a different networking library.

Patch-set: 5
Label: Code-Review=0