Update patch set 1
Patch Set 1: Code-Review-1

Hi Christopher! Thanks for taking the time to submit a pull request to the driver.

I can see where you're coming from with this. that being said, I don't think it's worth it to use a different capitalization in the string representation than we use in the name of the type itself merely because the shell capitalizes it that way. (In Go, it's preferred to capitalize the "D" in "ID" (see https://github.com/golang/go/wiki/CodeReviewComments#initialisms), which is why the type isn't called "ObjectId".) If the difference were large enough that it might be confusing to users, it might be worth it, but this isn't the case here.

Patch-set: 1
Label: Code-Review=-1