Add first draft implementation of connection type
Added the connection type implementation and just enough of the other
things required to bootstrap and initialize connection to the database
and complete the MongoDB Handshake.
Added implementation of a Pool with size and capacity.
Added internal implementation of fsm for service discovery and
monitoring.
Added server implementation for the topology package.
Added topology and server selection implementation for the topology
package.
Added dispatch package.
Removed dependencies on topology and connection packages from the
command package.

Change-Id: I8657939064d7279e7d2b8f23863dc3645d9d2b1b