Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> I'm not sure I understand the logic behind having a "description" package. This seems almost identical to the old "model" package, which we previously seemed to agree that we didn't want to have in the rewrite. Is this no longer the case?

The main issue I had with the model package was the name which implied it was a dumping ground for types. In reality, that's not what it is and it does serve a purpose. The description package removes a large part of what was in the model package (the fsm), and only has descriptions and the associated information.

Patch-set: 1