Handle nil properly in BSON encoder
Added support for nil interfaces, maps, pointers, and slices.
Added support for round tripping empty maps and slices and nil maps and
slices. Since these are distinct in Go, we use BSON Null for a nil map
or slice and an empty document or array, respectively, for zero value
map and slice types.

GODRIVER-317
GODRIVER-315

Change-Id: I96036138acb745b1f42d3190e9664366bb85a9a6