Update patch set 18
Patch Set 18:

(29 comments)

Concerning unit tests, we absolutely need unit test coverage. I was mainly focused on getting this initial refactor done and after we had settled on the design, then doing the unit tests. We can do those after this patch is applied.

There is another task of fixing up the integration tests and testing framework. We do have an epic for this, GODRIVER-251.

I haven't thought about OP_COMPRESS yet, but your suggestion of putting it inside the connection seems reasonable to me.

I made several tickets to handle the work of updating this patch. I did a large refactor of the library structure yesterday and I'd like to not try and reapply that plus the other tickets I did on top of this changing patch. For the few changes that don't have tickets I have another patch coming to fix those.

If no one objects, I say we merge this patch and handle the feedback in the subsequent tickets.

Patch-set: 18