Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> seems fine to me. Are we concerned that we are making it difficult for people underneath 1.9 to use the new driver? Is that not an issue? What about users converting from mgo that can't (for some reason) upgrade?

For the most part, I would like to use some of the features of Go 1.9 and 1.10 and try and track supported versions of the language. I have actually had to change some code and write (potentially buggy) code to backfill for features supported in 1.9 and 1.10 and I would like to stop doing that.

Patch-set: 2