Update patch set 2
Patch Set 2:

(10 comments)

My replies are below. The majority of complicated life cycle comes from the disconnection logic, mainly around ensuring that we've actually closed everything down. The difficulty for reconnection is mostly around not closing channels or making variables nil. I think it'll be useful enough in edge cases to warrant the small amount of variable recycling it requires.

The main issue is the connection pool which has some of the more complex logic. I think the race condition around Get can be solved with an additional check, inside the lock, if the pool has been disconnected. In the worst case we'll have dialed an additional connection and immediately close it.

The larger issue is that the design of the pool can currently lead to concurrent use of connections without knowing it if the user calls Close twice for a connection. I suspect that this will be a little more difficult to handle. There is also the issue that a user can Close a Connection, but continue to use it for Read and Write calls and receive no errors. This can lead to subtle bugs that might not be found until an application is under load and is doing a large number of quick operations.

Patch-set: 2
Label: Code-Review=0