Update patch set 1
Patch Set 1:

(3 comments)

Other than the notes below, the ticket seems to indicate a way to get the collection name AND database name from a collection, but all we have here is a way to get the collection name.

I would think providing a Database() method would be sufficient, because it has a Name() method on it. Kris may want something different, a new struct containing both together perhaps. If that's the case, I'd probably call this FullName() or Namespace().

Patch-set: 1
Label: Verified=0