Update patch set 1
Patch Set 1:

I think this implementation is right, Sam.

1) I have a question however, probably for Kris. This hook seems wildly inefficient. If someone wants to use this to invoke MGO's bson layer (which is the stated purpose), then it will require them convert to a []byte first, followed by converting from that []byte into a bson.Document, only for us to go back to a []byte to stick in on the wire.

So, initial reaction is why can this not just return a bson.Reader (or a []byte). I understand the current mechanism will do this inefficiently underneath due to how the bson package is designed, but that is something that can be changed later without needing the user to alter their usage.

2) The aggregate method takes an interface{}. However, it only truly accepts bson.Array, []*bson.Document, or an []interface{} (or defaulting to custom TransformDocument). However, if the user provides an old mgo []bson.D, this will completely fail, which seems to miss the point of allowing arbitrary transformation. In other words, there is a hole here.

Is there a particular reason we want to accept an interface{} instead of []interface{}? Changing the signature would fix the hole. I can't think of a reason a user would want to provide a document here.

Patch-set: 1
Reviewer: Craig Wilson <1013853@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0