Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I think this implementation is right, Sam.
> 
> 1) I have a question however, probably for Kris. This hook seems wildly inefficient. If someone wants to use this to invoke MGO's bson layer (which is the stated purpose), then it will require them convert to a []byte first, followed by converting from that []byte into a bson.Document, only for us to go back to a []byte to stick in on the wire.

I saw at least one case where we introspect bson.Document that TransformDocument returns, which might be awkward in the case that we have a []byte rather than a Document. A cursory investigation into a few other cases makes it seem like we're processing these results at least a little bit. 

> So, initial reaction is why can this not just return a bson.Reader (or a []byte). I understand the current mechanism will do this inefficiently underneath due to how the bson package is designed, but that is something that can be changed later without needing the user to alter their usage.

I think I was initially confused by the difference between bson.Reader and bson.Document, and I'm not sure that the above helps much. Introspecting the Readers is going to be awkward (and seems like it would be inefficient in some cases, but I'm not sure,) so it seems like we should do some profiling/performance goal setting outside of the scope of this ticket.

> 2) The aggregate method takes an interface{}. However, it only truly accepts bson.Array, []*bson.Document, or an []interface{} (or defaulting to custom TransformDocument). However, if the user provides an old mgo []bson.D, this will completely fail, which seems to miss the point of allowing arbitrary transformation. In other words, there is a hole here.

I think with a custom transformer, this would handle bson.D's correctly, but perhaps I'm missing something (I think that bson.D's are generally used for index specs and that's about it, and while I think providing a migration path is good, I'm also totally fine if bson.D is the casualty.) 

> Is there a particular reason we want to accept an interface{} instead of []interface{}? Changing the signature would fix the hole. I can't think of a reason a user would want to provide a document here.

Slices don't type convert to other slices without iterating through them, which is awkward from an application perspective. (Which is to say, while you can cast an interface{} to another type (e.g. bson.D/M/etc.), to cast a []interface{} to a slice of another type, you have to cast each item.)

Patch-set: 1
Label: Code-Review=0