Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Kris, 
> 
> After our conversation in person today (which resulted in generally deciding to have TransformDocument to return []byte, I want to know how we should handle cases where the input document would be types from our bson package? Should the plugable transform document only actually cover the reflection case?
> 
> I'll hold off on making any more changes until we have an idea of how we want things. 
> 
> Cheers,
> sam

We spoke again after I wrote this and decided that we should: handle type that we know about (e.g. the bson.Document and known types) directly, and use the configurable transformer to convert the existing "default" case. 

The problem here is that for the default implementation (and the case where its not a shim for the legacy bson library (which by volume is likely to be far more common) we'll end up creating bson.Document objects, converting them to byte slice, and then converting them back to bson.Document objects for processing

This puts me back in the camp of thinking that we should do something more like my initial process. The only downside is that it means that people who need to shim in different bson libraries will need to depend on our bson library for the converter. But these would be for shims, and I don't think it makes sense to make the library less reasonable for the common case in favor of supporting the shim case. 

I also don't think it makes sense to obscure the fact that the driver is using bson.Document objects internally. If we decide we want to change this later, I think it's also reasonable to push people toward using a different shim case.

Patch-set: 2
Reviewer: Sam Kleinman <1014178@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0