Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> > Patch Set 10:
> > 
> > (1 comment)
> > 
> > Concerning the WriteConcernError and the WriteError, since it is possible for a WriteConcernError to invalidate a WriteError, but the inverse is not true, we should return the WriteConcernError over the WriteError.
> 
> Could you elaborate on the "invalidated" write error?
> 
> Making fire-and-forget writes harder seems fine. (I expect in this case you'll just assign the err value to _, which is probably right?) but users can be left to their own devices.

Say you do two writes with a wtimeout and a w: majority, and the two writes have the same key. Let's further say that the primary cannot communicate with the secondaries, but the secondaries can all communicate with each other.

When the wtimeout expires, the first write will return with no write errors and a write concern error, the second write will return with a write error of duplicate key and a write concern error.

Let's further say that there is an election, a new primary is elected, and the old primary starts communicating with the rest of the replicaset. The entire timeline that caused the duplicate key error is now erased. Sending the same command again to the new primary will not result in a duplicate key error, making the duplicate key error a transitive error that is invalidated by the write concern error.

Patch-set: 10
Reviewer: Kris Brandow <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>