Update patch set 18
Patch Set 18: Code-Review-1

The test is still taking too long. We were already at around ~9 seconds for a test run of the mongo package. This brings that up to around ~30 seconds. With the race detector on, it's takes ~100 seconds to run when it only took ~10 seconds before. Since we run each package's test in both modes, this means that the tests are a minute and a half or more to run.

The amount of time it takes the tests to run was already too high, so we either need to vastly reduce the amount of time it takes this test to run, or introduce a flag for long running tests and only run this in evergreen.

Patch-set: 18
Reviewer: Kris Brandow <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1