Update patch set 18
Patch Set 18:

> From experience, if the tests take to long to run, people won't run them locally, which just pushes the problems later down the road. I want to eventually have an integration package in the mongo directory, similar to the integration package in core. I think this test would be a good candidate for that, possibly behind a flag for exhaustive tests.

I definitely sympathize with this experience and think that "the tests take too long" is difficult to fight and its important that we avoid test bloat. At the same time, having tests that only run in CI, or that are difficult to remember how to invoke outside of CI leads to a different kind of friction that also makes local development difficult and confusing for people with less access or familiarity with our CI tooling. 

I've opened a task (GODRIVER-425) to cover the work to reorganize integration tests, and in the mean time, I've added a skip for this test outside of evergreen.

Patch-set: 18
Reviewer: Sam Kleinman <1014178@d5d70762-12d0-45a1-890d-524b12d3f735>