Update patch set 20
Patch Set 20:

(1 comment)

> > I've opened a task (GODRIVER-425) to cover the work to reorganize integration tests, and in the mean time, I've added a skip for this test outside of evergreen.
> 
> I also would like to not have tests that we only run in CI. I think we'll build up a series of longer running tests (like fuzz tests) overtime, and we'll need a good way of running those. It might just be that we have a "-long" flag instead of a "-short" flag, and we tell people to run the -long flag before the make a commit.

I think this is isomorphic with just changing the default to short=true and adding a long argument, but I think that's an implementation detail for GODRIVER-425 and we shouldn't linger longer on this ticket.

I'm not going to rerun the patch, as my last change only edits a comment.

Patch-set: 20
Reviewer: Sam Kleinman <1014178@d5d70762-12d0-45a1-890d-524b12d3f735>