Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (1 comment)
> 
> Sorry I didn't notice this earlier, just occurred to me.

That makes sense. What about when the message is non-empty but the wrapped error is nil (e.g line 111 in scramsha1.go)? Would it make sense to update Error() in auth.go to just print the message if e.inner is nil or to go through and create an error so e.inner is never nil?

Patch-set: 4
Reviewer: Divjot Arora <1014410@d5d70762-12d0-45a1-890d-524b12d3f735>