Update patch set 3
Patch Set 3:

(2 comments)

we provide two ways to construct an array iterator (either calling the iterator() method on an existing array, or using the public constructor. I don't understand the usecase for having both, and I'm not clear which I should use as a developer.

I'd be in favor of removing one of these choices, but if that's not desirable, then we should document proper usage.

Patch-set: 3
Reviewer: Sam Kleinman <1014178@d5d70762-12d0-45a1-890d-524b12d3f735>