Update patch set 1
Patch Set 1:

So this issue is actually about tracking the internal state of the driver. We need to know if another goroutine has called cursor.Close. If it has, then when we can Next/getMore we should instead attempt to kill the cursor on the server and then exit.

We'll likely need to have a local variable on the cursor that we check to ensure that it hasn't been closed.

Patch-set: 1
Reviewer: Kris Brandow <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>