Update patch set 7
Patch Set 7: Code-Review+2

I think this is good for now. There are some of the options that we aren't printing the value but that's understandable because they would likely be unwieldy. I think we can merge this and if users have trouble debugging options in the future we can revisit how to better make debugging output.

Patch-set: 7
Reviewer: Kris Brandow <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2