Vendor and remove from internal JSON packages
The jsonparser and jsonpretty packages were placed in bson/internal.
There were not changes necessary to jsonpretty and the only change to
jsonparser can be implemented in the bson package using a closure. This
commit moves the libraries to the vendor directory, uses the fully copy
of the library, and updates the necessary code.

GODRIVER-480
GODRIVER-481

Change-Id: I5594ab0e753b27a4a18d959ec047341d988d3876