Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

I think it should be ClientSession. When a user specifies a session as an option, they only have the ClientSession to work with. Also, the server session field in Client is private as per the spec, so it can't be extracted in the mongo package.

Patch-set: 1
Reviewer: Divjot Arora <1014410@d5d70762-12d0-45a1-890d-524b12d3f735>