Update patch set 3
Patch Set 3:

> Patch Set 1:
> 
> This basically creates a Session struct that can be passed around as an option  to any CRUD command.  I didn't include DB/Collection/Client/Db opts (I don't see why they would)? The plan is for mongo.Session to wrap a core.ClientSession object. Let me know if this is going in the right direction.

So I think this is a good start. It seems that on the other side of Unbundle we will just take the session out and then attach it to the session property of a command, right?

Patch-set: 3
Reviewer: Kris Brandow <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>