Update patch set 12
Patch Set 12:

For patch set 12, we removed ReadConcern, WriteConcern, and ReadPreference as options from core altogether. If a user wants to have a different value for one of these for a command, they can use the Copy* helpers in collection.go to make a copy of the collection.

Patch-set: 12
Reviewer: Divjot Arora <1014410@d5d70762-12d0-45a1-890d-524b12d3f735>