Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I guess my only concern here is the same concern I had about read concern, write concern, and read preference, where we could have a different option in the options array than attached to the command. I think it's fine for session to be an option.Optioner, but maybe we should change the Unbundle methods to return option.Optioner always and in the driver, loop over the options and separate out the Session option from the other options.

I was just thinking about this. It might be a good idea to separate the session option in the collection.go methods, but this doesn't really stop Session from being passed as an option directly by someone not using the mongo package. For WC/RC/RP, we got around this by removing those options completely, but that doesn't seem possible for sessions.

Patch-set: 4