Update patch set 6
Patch Set 6:

We plan on changing the mongo session to hold a struct containing an updatable clock as discussed with Jeff so that we avoid propogating clustertime up through all the commands.  The rest of the code is up for review.

Patch-set: 6
Reviewer: Roland Fong <1014457@d5d70762-12d0-45a1-890d-524b12d3f735>