Update patch set 10
Patch Set 10:

> Patch Set 9:
> 
> The latest patch has a couple of logic updates I originally missed. For OP_QUERY, if a read pref is added, the query is wrapped in $query, so getting the actual command doc for CM needed a little extra logic. Also, the tests were timing out for all topologies when using auth and ssl. I think this was because the channels in the tests were being filled with auth messages and future channel writes were hanging. Running a new patch now, but increasing channel size seems to fix it so far.

For the OP_QUERY change, is there a test that fails if this new conditional is removed?  If not, can you add one?

For the channels that are filling up, it should not be due to auth messages.  Commands sent either for connection handshake or SDAM should not be reported, and if they are then that should be addressed.

Patch-set: 10
Reviewer: Jeff Yemin <1013848@d5d70762-12d0-45a1-890d-524b12d3f735>