Update patch set 6
Patch Set 6:

Update - wrote a quick test on the master branch to see if this error had to do with sessions specifically. The test inserts 3 documents and runs a find with various read prefs. I got the same error (couldn't find host matching read pref secondary). I was only running a patch for Mac OS 10.12, but it seems like the error happens for all combinations of auth/ssl/compression. However, I still can't recreate this error locally (running a sharded cluster with 3 shards, each of which is a replica set with 1 primary and 2 secondaries. The cluster has 1 mongos and 1 config server).

I thought it was strange that the test passes on versions <3.6, so I thought it might be related to OP_MSG. I reset master to fff8c98e325fa0c123220662024befb893f480bf (after options management but before the OP_MSG work was merged in) and the test still doesn't pass for the same reason.

Also, the test plan for list collections requires replica sets to be tested with primary and secondary read prefs, but only requires sharded clusters to be tested with primary. Any reason why this is? Is secondary read pref even valid for sharded clusters?

Patch-set: 6
Reviewer: Divjot Arora <1014410@d5d70762-12d0-45a1-890d-524b12d3f735>