Update patch set 2
Patch Set 2:

(15 comments)

Response to comments are below.

The reason that the *Value and *Element modes exist is because I found it difficult to reason about what was happening by attempting to use other heuristics to determine what mode one is actually in. For instance, we could add a "onElement" and an "onValue" property to the stack frame, but at the point you are adding a stack frame without actually adding a stack frame. This feels inconsistent. Instead, we should use an actual stack frame where it makes sense and I think when we move from a Document to an Element and from an Array to a Value we should add an additional stack. The only reason I can see for not doing it that way and for adding the information to the parent stack is to avoid allocations from pushing and popping. The stack implementation from the reader solves this problem by switching from a stack that always allocates to sliding along a slice of stack frames. This is likely even better since the stack frames are a values so they are located in close proximity in memory as well. The number of allocations we'll need to do isn't too high either, since we'll be doubling the size of the slice each time, as opposed to the current implementation for the value and document writer where we allocate a new stack frame for each depth level (and for each element/value).

This does introduce the need to jump two stack frames when popping, but that's solved in the reader implementation by just having the pop method always pop 2 stack frames, which is really just decrementing an int and copying the offset.

Ultimately, I don't see the value in checking for two values (the mode and the "onElement" or "onValue" property) when I can just check for one. Furthermore, checking for two values means it's easier for a maintainer to forget to check one of them, or to check the wrong one. This is really difficult logic to understand, and only using the stack machine for depth confused me quite a bit at first. If we are going to use a stack, I would like to use the stack for everything.

Patch-set: 2
Reviewer: Kris Brandow <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>