Update patch set 1
Patch Set 1:

This feels way to high to be checking for a document being oversized... The lower level thing that writes document onto the wire should be handling that.

This is fine as a quick fix patch, but it doesn't fix the general problem. This same thing could happen with updates or queries or deletes or ... Should we check in all those places?

Patch-set: 1
Reviewer: Craig Wilson <1013853@d5d70762-12d0-45a1-890d-524b12d3f735>