Update patch set 5
Patch Set 5:

I'm fine with the code... But a question here: Do we really want to take a dependency for something like this? It'd be one thing if this were an application, but we are an upstream publisher. It means we are requiring all of our users to take this dependency as well. And this is a pretty small thing for a dependency?

I see a lot of go libraries out there indicating "zero-dependencies" and I, as a consumer, find a lot of value in that.

Patch-set: 5
Reviewer: Craig Wilson <1013853@d5d70762-12d0-45a1-890d-524b12d3f735>