Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> SASLprep requires the x/text/* libraries anyway.  Once you take a dependency for part of the solution, is it that much worse to depend on the rest?

Well, we can differentiate between go's core libraries (I know x isn't in the stdlib directly, but...) and something else.

But even then, the question becomes what is being depended on in x/text and why that dependency is required? I'm really not a NIH guy, but libraries are just different. In .NET taking a dependency in a library has to be thought about really hard. Different languages require different things, so perhaps consider this more of a question for Kris...

Is taking a dependency in golang (in a redistributable library) a big deal? Under what conditions is it not a problem (or is problem), etc...

Patch-set: 5
Reviewer: Craig Wilson <1013853@d5d70762-12d0-45a1-890d-524b12d3f735>