Update patch set 6
Patch Set 6:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > SASLprep requires the x/text/* libraries anyway.  Once you take a dependency for part of the solution, is it that much worse to depend on the rest?
> 
> Well, we can differentiate between go's core libraries (I know x isn't in the stdlib directly, but...) and something else.
> 
> But even then, the question becomes what is being depended on in x/text and why that dependency is required? I'm really not a NIH guy, but libraries are just different. In .NET taking a dependency in a library has to be thought about really hard. Different languages require different things, so perhaps consider this more of a question for Kris...
> 
> Is taking a dependency in golang (in a redistributable library) a big deal? Under what conditions is it not a problem (or is problem), etc...

Historically, the goal in Go was to always have the fewest possible number of dependencies possible. In many cases this meant opting for actually copying in code instead of creating a dependency. I would like to keep the number of dependencies down.

I realize that the x/text/* packages exist, but is there a way we can do this with the unicode package in the standard library?

I think it's also less of a problem to have dependencies in Go now that we have dep and Go 1.11 modules, but I would still very much like to reduce the number of dependencies as much as possible.

Patch-set: 6
Reviewer: Kris Brandow <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>