Update patch set 2
Patch Set 2:

The one thing I'm missing is a simple view of how a client would interact with this? It appears to be just a cursor, using the same implementation that exists internally for other things (ss.BuildCursor).

In particular, if the server sends no data, but doesn't close the cursor, what happens? This has come up before in other conversations and I don't remember where we landed or what the behavior is, although I suspect it's not what we want.

Patch-set: 2
Reviewer: Craig Wilson <1013853@d5d70762-12d0-45a1-890d-524b12d3f735>