Update patch set 1
Patch Set 1:

This looks good, but let's add some tests to ensure that we are actually properly sending the maxTimeMS on the getMore commands. Also, although the spec doesn't say it, maxAwaitTime is a valid option for Aggregate, so that should be added and then the actual maxAwaitTime can be passed to aggregate.

Patch-set: 1
Reviewer: Kris Brandow <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>