Update patch set 1
Patch Set 1:

(1 comment)

So I made a comment on the code, but now that I think about it more, I don't think this is the right approach to fix this. I'm thinking of a case where a user has a custom codec and they want to control how a BSON Null is handled, e.g. maybe a user wants to set a default or perhaps they want to set another field. I think instead we should add support for the handling BSON Null to all of the default and primitive decoders (and add support for encoding to BSON Null for pointers in the default and primitive encoders). This means the struct codec won't arbitrarily take away command from the codec process, but it also means that BSON Null will be properly handled for maps, slices, and arrays.

Patch-set: 1
Reviewer: Kris Brandow <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>