Redesign Document, Array, Element, and Value types
Migration Notes:
The Document, Element, Value, and Array types have been redesign and
rewritten.

The Value type no longer stores values as a slice of bytes,
instead opting for a compact 32byte representation that stores values
closer to native Go types. An empty Value is equivalent to a BSON Null.
All of the methods in this package will treat Value{} and
bson.Null(bson.NullPrimitive{}) as the same.

The Element type is much simpler and adds a key to a Value.

The Document and Array types have been simplified, removing many of
their methods, and making them conceptually two different types, instead
of building an Array on top of a Document. Their methods have also been
updated to handle easier usage of Element and Value instances.

The constructors have been removed and replaced with a simpler
construction system that uses functions. There is a symmetry between
these functions and the methods on the Value type. The input to the
constructor matches the output of the Value method, for example this is
valid: bson.Double(val.Double()).

GODRIVER-413

Change-Id: I353a83b7c8f8c5fa0c2552de9ff8bc78d55c50bc