Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> yup, looks great.
> 
> My last concern I think, is still that Embed method, which is used to turn a Document/Array into a Value, not necessarily embed them in something. I don't know what to call this method, DocumentValue? ArrayValue?
> 
> Also, we should should have a method that takes an interface{} and turns it into a Value as well. ToValue?

Embed is the right thing for document, I'm not sure if it's the right thing for Array, but you are, in effect, embedding the Document or Array into the Value.

I don't particularly want to have an interface{} constructor because it won't fit with the rest of the constructors and I was thinking that Document/Array/Element/Value form a closed system. Perhaps we should wait and see if anyone needs it, because implementing it will likely require another ValueWriter implementation, or to go through a []byte.

Patch-set: 7
Reviewer: Kris Brandow <1001503@d5d70762-12d0-45a1-890d-524b12d3f735>