Update patch set 9
Patch Set 9:

(2 comments)

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > yup, looks great.
> > 
> > My last concern I think, is still that Embed method, which is used to turn a Document/Array into a Value, not necessarily embed them in something. I don't know what to call this method, DocumentValue? ArrayValue?
> > 
> > Also, we should should have a method that takes an interface{} and turns it into a Value as well. ToValue?
> 
> Embed is the right thing for document, I'm not sure if it's the right thing for Array, but you are, in effect, embedding the Document or Array into the Value.
> 
> I don't particularly want to have an interface{} constructor because it won't fit with the rest of the constructors and I was thinking that Document/Array/Element/Value form a closed system. Perhaps we should wait and see if anyone needs it, because implementing it will likely require another ValueWriter implementation, or to go through a []byte.

Embed is the right thing for a document when you are embedding it into another container. Same with any type really. The problem is that Embed is the way to turn a Document/Array into a Value, which might not always be for embedding purposes. This is true in Mongohouse in a number of places. 

So we should have another name for this other than Embed... something that turns a Document into a value, the same way we have Int32 to turn an int32 into a value or Bool to turn a bool into a value, both ready to "embed" into a container.

Patch-set: 9
Reviewer: Craig Wilson <1013853@d5d70762-12d0-45a1-890d-524b12d3f735>