Update patch set 14
Patch Set 14:

I'd truly encourage you to do the benchmarking as part of this ticket. We are already going to move everyone's cheese who has used this bson library at all. This is way more intrusive than simply renaming the parse extended json function names. I'm not relishing the amount of change this will cause for Mongohouse. What is very important here is to not introduce a new API only to find that it is slower and more memory deficient than bson.D and bson.M and force everyone who adopted it to adopt yet another API.

We need this to be final (in broad strokes) and the only way to ensure ourselves of that is to (a) ensure we like the API. I don't think that's a problem because we generally like the bson.D/bson.M API and (b) ensure that it is up to the task performance wise.

Patch-set: 14
Reviewer: Craig Wilson <1013853@d5d70762-12d0-45a1-890d-524b12d3f735>